refactor: refactored part of the codebase; identified a circular import issue that needs to be resolved.
