+++ src/handlers/__init__.py
@@ -1,9 +1,9 @@
-from .authorization import authorized_only
+from .authorization import authorized_only, temp_authorize_user_by_contact
 
 from .birthdays import send_birthdays_month, back_to_birthdays
 
 from .commands import send_main_menu, proceed_authorize_users, toggle_admin_mode, temp_authorize_user, \
-    send_mass_message,  remind_password, proceed_mass_message, old_button_handler
+    send_mass_message, remind_password, proceed_mass_message, old_button_handler
 
 from .departments import send_departments, send_inter_department_contacts, send_department_contacts, add_director, \
     proceed_add_director, send_sub_departments_contacts, send_search_form, proceed_contact_search, \
@@ -23,3 +23,9 @@ from .links import send_business_process, send_business_processes_menu, add_link
 
 from .main_menu import send_knowledge_base, send_business_processes, send_birthdays, send_contacts_menu, thanks_menu, \
     send_useful_links, send_form, back_to_send_contacts_menu
+
+from .commendations import show_thanks, show_my_thanks, show_thanks_period, show_commendation, delete_commendation, \
+    confirm_delete_commendation, send_thanks, proceed_thanks_search, proceed_send_thanks, send_thanks_name, \
+    confirm_send_thanks, com_change_position, com_change_position_ans, cancel_send_thanks, hide_message
+
+from .ai import ai_question, proceed_ai_question, cancel_ai_question
