+++ src/handlers/authorization.py
@@ -1,7 +1,7 @@
-from telebot import types
+from telebot import types, apihelper
 
 from src.database import DatabaseConnection
-from src.config import authorized_ids, bot
+from src.config import authorized_ids, bot, process_in_progress
 
 
 def authorized_only(user_type):
@@ -34,3 +34,33 @@ def authorized_only(user_type):
         return wrapper
 
     return decorator
+
+
+@bot.message_handler(func=lambda message: process_in_progress.get(message.chat.id) == 'temp_authorization',
+                     content_types=['contact'])
+@authorized_only(user_type='admins')
+def temp_authorize_user_by_contact(message):
+    new_user_id = message.contact.user_id
+    if new_user_id not in authorized_ids['users'] and new_user_id not in authorized_ids['temp_users']:
+        authorized_ids['temp_users'].add(new_user_id)
+
+        try:
+            bot.send_message(new_user_id, f'Вас тимчасово авторизовано адміністратором @{message.from_user.username}.')
+
+            log_text = (f'User {new_user_id} temporarily authorized by @{message.from_user.username} with notification.'
+                        f'\nTemporarily authorized users: {authorized_ids["temp_users"]}')
+        except apihelper.ApiTelegramException:
+            log_text = (
+                f'User {new_user_id} temporarily authorized by @{message.from_user.username} without notification.'
+                f'\nTemporarily authorized users: {authorized_ids["temp_users"]}')
+
+        print(log_text)
+
+        bot.send_message(message.chat.id, f'✅ Користувача <b>{message.contact.first_name}</b> авторизовано.',
+                         parse_mode='HTML')
+
+    else:
+        bot.send_message(message.chat.id, f'🚫 Помилка авторизації:'
+                                          f'\nКористувач <b>{message.contact.first_name}</b> вже авторизований.',
+                         parse_mode='HTML')
+    del process_in_progress[message.chat.id]
