+++ src/handlers/main_menu.py
@@ -4,9 +4,8 @@ from telebot import types
 
 from src.config import bot, MONTH_DICT, authorized_ids, make_card_data, process_in_progress
 from src.handlers import authorized_only
-from src.handlers.authorization import authorized_only
-from src.integrations.google_forms_filler import send_question_form
-from src.utils.messages import send_links
+from src.integrations import send_question_form
+from src.utils import send_links
 
 
 @bot.message_handler(func=lambda message: message.text == '🎓 Навчання')
