+++ src/integrations/__init__.py
@@ -1,3 +1,5 @@
 from .crm_api_functions import get_employee_pass_from_crm, add_employee_to_crm, delete_employee_from_crm, \
     update_employee_in_crm
 from .telethon_functions import proceed_find_user_id, send_photo, decrypt_session, remove_user_from_chat
+from .google_forms_filler import send_question_form
+from .google_api_functions import update_employees_in_sheet
