+++ src/integrations/google_forms_filler.py
@@ -12,7 +12,7 @@ from telebot import types
 
 from src.config import bot, user_data, process_in_progress
 from src.database import DatabaseConnection
-from src.utils.messages import delete_messages
+from src.utils import delete_messages
 
 
 class FormFiller:
