+++ src/utils/__init__.py
@@ -1,4 +1,4 @@
-from .make_card import make_card
-from .messages import delete_messages, hide_message
-from .reminder import scheduler
 from .main_menu_buttons import main_menu, button_names, old_button_names
+from .make_card import make_card
+from .messages import delete_messages, send_links
+from .scheduler import start_scheduler
